requirements.txt
certifi==2020.11.8
chardet==3.0.4
dataclasses-json==0.5.2
idna==2.10
isodate==0.6.0
marshmallow==3.9.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.0
python-dotenv==0.15.0
python-youtube==0.6.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
stringcase==1.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
youtube-dl==2020.12.2
